// src/interfaces/auth.ts
export interface User {
    id: number
    username: string
    email: string
    firstName: string
    lastName: string
    gender?: string
    image?: string
    accessToken?: string,
}

export interface LoginRequest {
    username: string
    password: string
}

export interface LoginResponse extends User{
    accessToken: string,
    refreshToken: string
}

export interface RefreshTokenResponse {
    accessToken: string
    refreshToken: string
}